renaming test files
