+++ example/kafka-1-kafkajs/app.ts
@@ -1,0 +1,11 @@
+import createKafka from './index'
+
+const main = async () => {
+  const kafka = createKafka()
+
+  return {
+    kafka,
+  }
+}
+
+export default main
