+++ example/kafka-1-kafkajs/kafka.d.ts
@@ -1,0 +1,1 @@
+declare module 'kafkajs'
