+++ example/postgres-2-knex/app.spec.ts
@@ -1,0 +1,15 @@
+import main from './app'
+// @ts-ignore
+import { seedBanana } from './data.json'
+
+describe('postgres-2-knex', () => {
+  it('trabajo', async () => {
+    const result = await main()
+
+    expect(result).toEqual(
+      expect.objectContaining({
+        firstEntry: expect.objectContaining(seedBanana),
+      })
+    )
+  })
+})
